+++ service_crategen/src/cargo.rs
@@ -30,6 +30,7 @@ pub struct Metadata {
     pub version: String,
     pub homepage: Option<String>,
     pub edition: String,
+    pub exclude: Option<Vec<String>>,
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize)]